refactor: move mainSectinSlice to store/slices
