+++ src/widgets/BhandaraCheckin/store/actions/resetAppState.ts
@@ -1,6 +1,6 @@
 import { AnyAction, Dispatch } from "@reduxjs/toolkit";
 import { bhandaraCheckinSlice } from "../../BhandaraCheckin/bhandaraCheckinSlice";
-import { mainSectionSlice } from "../../SectionMain/mainSectionSlice";
+import { mainSectionSlice } from "../slices/mainSectionSlice";
 import { updateDetailsSectionSlice } from "../slices/updateDetailsSectionSlice";
 
 export const resetAppState = (dispatch: Dispatch<AnyAction>) => {
